Replace onMouseOver with group-hover and refactor Code
