+++ app/people/_components/ContributorWrapper.tsx
@@ -1,11 +1,8 @@
-import { parseDateRangeSearchParam } from "@/lib/utils";
 import ContributorsPage from "./ContributorsPage";
 import { getContributorsData } from "@/app/api/leaderboard/functions";
-import { Contributor, ContributorsPageProps } from "@/lib/types";
+import { PageProps } from "@/lib/types";
 
-export default async function ContributorWrapper({
-  searchParams,
-}: ContributorsPageProps) {
+export default async function ContributorWrapper({ searchParams }: PageProps) {
   const data = await getContributorsData();
   return <ContributorsPage contributors={data} searchParams={searchParams} />;
 }
