+++ app/people/page.tsx
@@ -1,13 +1,18 @@
-import Searchbar from "./_components/Searchbar";
+import SearchBar from "@/components/SearchBar";
 import { Suspense } from "react";
 import ContributorWrapper from "./_components/ContributorWrapper";
-import { ContributorsPageProps } from "@/lib/types";
+import { PageProps } from "@/lib/types";
 import LoadingText from "@/components/LoadingText";
 
-export default async function Page({ searchParams }: ContributorsPageProps) {
+export default async function Page({ searchParams }: PageProps) {
   return (
     <div className="mx-auto mb-20 flex max-w-full flex-col items-center justify-center gap-8 px-24">
-      <Searchbar searchParams={searchParams} />
+      <SearchBar
+        searchParams={searchParams}
+        isDaterange={false}
+        isSort={false}
+        isRoleFilter={false}
+      />
       <Suspense fallback={<LoadingText text="Finding the contributors" />}>
         <ContributorWrapper searchParams={searchParams} />
       </Suspense>
